Load a glb object with pre-existing animation. Place all JS code in one folder.
