+++ Source/ui/public/index.html
@@ -12,7 +12,7 @@
           }
         }
     </script>
-    <script type="module" src="index.js"></script>
+    <script type="module" src="js/index.js"></script>
     <title>Reverbulizer</title>
     <link rel="icon" type="image/x-icon" href="/assets/favicon.ico">
     <link rel="stylesheet" href="index.css" >
