+++ Source/ui/public/js/juce/check_native_interop.js
